requirements_dicom_analyzer.txt
# Зависимости для DICOM Analyzer (Упрощенная версия)
# ===================================================

# Основные зависимости
torch>=2.0.0
transformers>=4.30.0
pydicom>=2.3.0
Pillow>=9.0.0
numpy>=1.21.0

# Дополнительные зависимости для лучшей работы
accelerate>=0.20.0
safetensors>=0.3.0

# Для прогресс-бара
tqdm>=4.64.0

# Для работы с CUDA (опционально)
# torch-audio  # раскомментировать если нужна поддержка аудио
# torch-vision  # раскомментировать если нужна поддержка vision
